+++ spec/datarockets_style/version_spec.rb
@@ -1,0 +1,7 @@
+RSpec.describe "DatarocketsStyle::Version" do
+  describe "::VERSION" do
+    it "contains the current version" do
+      expect(DatarocketsStyle::VERSION).not_to be nil
+    end
+  end
+end